+++ table/block_based_table_reader.h
@@ -118,7 +118,7 @@ class BlockBasedTable : public TableReader {
   // Returns all DataBlocks on SST File.
   Status GetDataBlocks(const ReadOptions&,
                        std::vector<char>& data,
-                       std::vector<uint32_t>& seek_indices) override;
+                       std::vector<uint64_t>& seek_indices) override;
 
   FragmentedRangeTombstoneIterator* NewRangeTombstoneIterator(
       const ReadOptions& read_options) override;
